Added static colors to pie chart Added Labels that are colors static (for legend)
Need looking to add next:
Still need to fix the methods to setBudgetMax (db not ready for it)
Still need to add onRestart() for mainActivity to repopulate tabs.

Issues:
There is a bug with the first draw of the pie graph - though when swiping through to end and back to start it fixes itself.